chore: Add DNS-Toolbox demo functions for NS, TXT, CNAME, and MX records
