+++ internal/dns/dns.py
@@ -1,0 +1,24 @@
+import dns.resolver
+
+#MARK: - DNS Servers
+DNS_SERVERS = {
+    "Google": "8.8.8.8",
+    "Cloudflare": "1.1.1.1",
+    "Quad9": "9.9.9.9"
+}
+
+#MARK: - DNS Function
+def get_dns_record(domain, dns_server, type):
+    resolver = dns.resolver.Resolver()
+    resolver.nameservers = [dns_server]
+    try:
+        answers = resolver.resolve(domain, type)
+        return [str(rdata) for rset in answers.response.answer for rdata in rset]
+    except resolver.NoAnswer:
+        return ["Error: No answer from the DNS server for the requested domain and record type."]
+    except resolver.NXDOMAIN:
+        return ["Error: The requested domain does not exist."]
+    except resolver.NoNameservers:
+        return ["Error: No nameservers are available to fulfill the request."]
+    except resolver.Timeout:
+        return ["Error: The DNS request timed out."]
